repo with JDBC template and all java classes moved to packages
