+++ src/main/java/com/elster/tacos/api/OrderController.java
@@ -1,5 +1,6 @@
-package com.elster.tacos;
+package com.elster.tacos.api;
 
+import com.elster.tacos.model.Order;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
