+++ src/main/java/com/elster/tacos/model/Ingredient.java
@@ -1,4 +1,4 @@
-package com.elster.tacos;
+package com.elster.tacos.model;
 
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
@@ -8,7 +8,7 @@ public class Ingredient {
     private final String id;
     private final String name;
     private final Type type;
-    public static enum Type {
+    public enum Type {
         WRAP, PROTEIN, VEGGIES, CHEESE, SAUCE
     }
 }
